Fix: CIF files with leading space characters now handled correctly.
This patch fixes a bug in the readCIF function that prevented it from correctly reading CIF files that contain leading space characters. The patch now removes any leading space characters before reading the file.

For example, the r_spositions function of cif.R now correctly searches for lines that begin with the character "_".